:lipstick: style : QA 피드백 수정
