+++ src/app/globals.css
@@ -102,8 +102,8 @@
 }
 
 .mySwiper2 {
-  height: 80%;
   width: 100%;
+  height: 590px;
 }
 
 .mySwiper {
