+++ src/constant.ts
@@ -1,3 +1,3 @@
-const SITE_URL = process.env.NEXT_PUBLIC_DOMAIN;
+const SITE_URL = process.env.NEXT_PUBLIC_BASE_URL;
 
 export default SITE_URL;
